+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * `JSON-dry` can now directly undry an object, no need to stringify it first
 * You can add a `interceptOpenUrl` method on the `Scene` class to intercept open urls
 * Add client-side connection checking
+* Add `use strict` to all helpers
+* Add `module.exports` headers to helpers automatically
 
 ## 1.2.2 (2018-01-05)